Shifted camera initialisation code into camera_interface
- Deleted temporary camera initialisation modules
