+++ RobotNoASF/Interfaces/camera_buffer_interface.c
@@ -32,7 +32,6 @@
 
 //////////////[Includes]////////////////////////////////////////////////////////////////////////////
 #include "../robot_setup.h"
-#include <stdio.h>				//sizeof()
 #include "camera_interface.h"
 #include "external_interrupt.h"	//External interrupts used to fetch frame from camera
 #include "timer_interface.h"	//Provides delay_ms()
@@ -79,8 +78,6 @@
 #define DO5				(DO5_PORT->PIO_PDSR & DO5_PIN)
 #define DO6				(DO6_PORT->PIO_PDSR & DO6_PIN)
 #define DO7				(DO7_PORT->PIO_PDSR & DO7_PIN)
-
-
 
 // Reading from buffer to SAM4
 #define readResetOn		RRST_PORT->PIO_CODR	|= RRST_PIN		//Buffer read reset
