--- RobotNoASF/Interfaces/camera_initialisation.h
@@ -1,33 +1,0 @@
-/*
-* camera_initialisation.h
-*
-* Author : Matthew Witt (wittsend86@gmail.com)
-* Created: 24/03/2018 1:02:26 PM
-*
-* Project Repository: https://github.com/wittsend/aut-swarm-robotics
-*
-* Header file for camera initialisaition module
-*
-* More Info:
-* Atmel SAM 4N Processor Datasheet:http://www.atmel.com/Images/Atmel-11158-32-bit%20Cortex-M4-Microcontroller-SAM4N16-SAM4N8_Datasheet.pdf
-* Relevant reference materials or datasheets if applicable
-*
-* Functions:
-* void funcName(void)
-*
-*/
-
-
-#ifndef CAMERA_INITIALISATION_H_
-#define CAMERA_INITIALISATION_H_
-
-//////////////[Includes]////////////////////////////////////////////////////////////////////////////
-
-//////////////[Public Defines]//////////////////////////////////////////////////////////////////////
-
-//////////////[External Global Variables]///////////////////////////////////////////////////////////
-
-//////////////[Functions]///////////////////////////////////////////////////////////////////////////
-uint16_t camSetup2(void);
-
-#endif /* CAMERA_INITIALISATION_H_ */
