+++ RobotNoASF/swarm_robot_V2.cproj
@@ -272,12 +272,6 @@
     <Compile Include="Interfaces\camera_buffer_interface.h">
       <SubType>compile</SubType>
     </Compile>
-    <Compile Include="Interfaces\camera_initialisation.c">
-      <SubType>compile</SubType>
-    </Compile>
-    <Compile Include="Interfaces\camera_initialisation.h">
-      <SubType>compile</SubType>
-    </Compile>
     <Compile Include="Interfaces\camera_interface.c">
       <SubType>compile</SubType>
     </Compile>
